requirements.txt
numpy
opencv-python-headless
mediapipe
streamlit
streamlit_webrtc
kora
pyvideoreader
